% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_text.R
\name{generate_text}
\alias{generate_text}
\title{Generate random text using Markov chains.}
\usage{
generate_text(word_freqs, word_length = 200, start_word = NA, rnd_seed = NA)
}
\arguments{
\item{word_freqs}{A \code{tbl_df} output from \code{markov_text::get_word_freqs()}.}

\item{word_length}{(Default 200) The length in words of the text to generate.}

\item{start_word}{(Optional) A starting word.}

\item{rnd_seed}{(Optional) A random number seed. Useful for reproducing results.}
}
\value{
A character vector containing randomly generated text.
}
\description{
Generate random text using Markov chains.
}
\examples{
\dontrun{wordfreqs_prideprejudice <- get_word_freqs(dplyr::tibble(text = janeaustenr::prideprejudice))
generate_text(wordfreqs_prideprejudice)
}
}
