% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_text.R
\name{get_word_freqs}
\alias{get_word_freqs}
\title{Generate word-frequency table for Markov text generation.}
\usage{
get_word_freqs(text_input, num_words = 500, n_grams = c(2, 3))
}
\arguments{
\item{text_input}{A \code{tbl_df} with a column named \code{text}.}

\item{num_words}{(Default 500) The number of unique words to include in the table.}

\item{n_grams}{(Default 2) Accepts values 2 or 3. How many words should we look at when computing frequencies?}
}
\value{
A \code{tbl_df} suitable for markov_text::generate_text().
}
\description{
Generate word-frequency table for Markov text generation.
}
\examples{
\dontrun{wordfreqs_prideprejudice <- get_word_freqs(dplyr::tibble(text = janeaustenr::prideprejudice))
generate_text(wordfreqs_prideprejudice)
}
}
