% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wordfreqs_dougford_2grams}
\alias{wordfreqs_dougford_2grams}
\title{2-Gram word-frequency Markov table for Doug Ford's victory speech.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 735 rows and 4 columns.
}
\source{
\url{https://www.macleans.ca/politics/doug-fords-victory-speech-ontario-is-open-for-business-full-transcript/}
}
\usage{
wordfreqs_dougford_2grams
}
\description{
For use with the function \code{markovtext::generate_text()}.
}
\keyword{datasets}
