% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wordfreqs_onefishtwofish_3grams}
\alias{wordfreqs_onefishtwofish_3grams}
\title{3-Gram word-frequency Markov table for One Fish, Two Fish, Red Fish, Blue Fish by Dr. Seuss.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1326 rows and 5 columns.
}
\usage{
wordfreqs_onefishtwofish_3grams
}
\description{
For use with the function \code{markovtext::generate_text()}.
}
\keyword{datasets}
