% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wordfreqs_whale_3grams}
\alias{wordfreqs_whale_3grams}
\title{3-Gram word-frequency Markov table for the whale's monologue from The Hitchhiker's Guide to the Galaxy by Douglas Adams.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 174 rows and 5 columns.
}
\usage{
wordfreqs_whale_3grams
}
\description{
For use with the function \code{markovtext::generate_text()}.
}
\keyword{datasets}
