% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wordfreqs_zarathustra_3grams}
\alias{wordfreqs_zarathustra_3grams}
\title{3-Gram word-frequency Markov table for the first four chapters of Thus Spake Zarathustra by Friedrich Nietzsche.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2106 rows and 5 columns.
}
\source{
\url{https://www.gutenberg.org/files/1998/1998-h/1998-h.htm#link2H_4_0004}
}
\usage{
wordfreqs_zarathustra_3grams
}
\description{
For use with the function \code{markovtext::generate_text()}.
}
\keyword{datasets}
